chore: Dump all WIP code
These changes include a basic templating engine powered by Netlify CMS. To
access the console, go to localhost:8000/admin. After authenticating, you must
copy the URL fragment (everything after utd.guide/) and go to
localhost:8000/<URL_FRAGMENT> Content is stored in the content/ directory. See
the existing files for a template. This code needs a lot of work. Do not merge
it directly to master or a develop branch.