fix: Resolve build errors
The site didn't properly check for the window variable which wasn't available
during Gatsby's server-side rendering. Additionally, the site didn't build due
to a missing icon for the gatsby-plugin-manfiest plugin. There was also an issue
with a missing page query for the profile page and a broken WIP FAQ page that
can be removed for now.