Finished talwind.config.js
I finished changing the colors and font styles according to the given information but there are certain bugs in this for example the line height is not right as often multiple text elements seem to touch each other and also primary-blue-5 and neutral-5 seem to be the same. I don't know if I should delete the previous code or comment it out so I just deleted it.